requirements.txt
pyautogui==0.9.53
opencv-python==4.6.0.66
Pillow==9.2.0
numpy==1.23.3
pytesseract==0.3.10
openai==0.27.7
requests==2.28.1
python-dotenv==0.21.0 
customtkinter==5.2.0 